Safari-611.1.10/LayoutTests/accessibility/ios-simulator/element-in-table-cell.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body id="body">

<table>
    <thead>
        <tr>
            <th scope="col">apple</th>
            <th scope="col">orange</th>
        </TR>
    </thead>
    <tbody>
        <tr>
            <td><button id="button1">3</button></td>
            <td><a href="#" id="link1">5</a></td>
        </TR>
    </TBODY>
</table>

<button id="button2">2</button>
<a href="#" id="link2">4</a>

<p id="description"></p>
<div id="console"></div>

<script>
    description("This tests the _accessibilityIsInTableCell method, used by VO to describe the element.");

    if (window.accessibilityController) {
        shouldBeTrue("accessibilityController.accessibleElementById('button1').isInTableCell()")
        shouldBeFalse("accessibilityController.accessibleElementById('button2').isInTableCell()")
        shouldBeTrue("accessibilityController.accessibleElementById('link1').isInTableCell()")
        shouldBeFalse("accessibilityController.accessibleElementById('link2').isInTableCell()")
    }
</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
