LayoutTests/accessibility/ios-simulator/element-in-table-cell-expected.txt

apple	orange
3	5
2 4
This tests the _accessibilityIsInTableCell method, used by VO to describe the element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibilityController.accessibleElementById('button1').isInTableCell() is true
PASS accessibilityController.accessibleElementById('button2').isInTableCell() is false
PASS accessibilityController.accessibleElementById('link1').isInTableCell() is true
PASS accessibilityController.accessibleElementById('link2').isInTableCell() is false
PASS successfullyParsed is true

TEST COMPLETE
